backend/requirements.txt
Django==4.0
djangorestframework==3.13.1
django-allauth==0.44.0
